requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.7
django-environ==0.11.2
djangorestframework==3.14.0
idna==3.4
numpy==1.26.2
pandas==2.1.3
pybitflyer==0.1.9
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
